split value classes to separate files
