+++ packages/squiggle-lang/src/library/index.ts
@@ -3,7 +3,7 @@ import { INDEX_LOOKUP_FUNCTION } from "../expression/constants.js";
 import { BuiltinLambda, Lambda } from "../reducer/lambda.js";
 import { Bindings } from "../reducer/stack.js";
 import { ImmutableMap } from "../utility/immutableMap.js";
-import { vLambda } from "../value/index.js";
+import { vLambda } from "../value/vLambda.js";
 import { makeMathConstants } from "./math.js";
 import { makeDefinition } from "./registry/fnDefinition.js";
 import { frAny } from "./registry/frTypes.js";
