+++ packages/squiggle-lang/src/library/version.ts
@@ -1,5 +1,6 @@
 import { ImmutableMap } from "../utility/immutableMap.js";
-import { Value, vString } from "../value/index.js";
+import { Value } from "../value/index.js";
+import { vString } from "../value/VString.js";
 
 // automatically updated on release by ops/ patch-js utils
 const VERSION = "0.9.3-0";
