+++ packages/squiggle-lang/src/public/SqProject/index.ts
@@ -6,7 +6,7 @@ import { createContext } from "../../reducer/context.js";
 import { Bindings } from "../../reducer/stack.js";
 import { ImmutableMap } from "../../utility/immutableMap.js";
 import * as Result from "../../utility/result.js";
-import { vDict, VDict } from "../../value/index.js";
+import { vDict, VDict } from "../../value/VDict.js";
 import { SqError, SqOtherError } from "../SqError.js";
 import { SqLinker } from "../SqLinker.js";
 import { SqValue, wrapValue } from "../SqValue/index.js";
