+++ packages/squiggle-lang/src/value/VVoid.ts
@@ -1,0 +1,13 @@
+import { BaseValue } from "./index.js";
+
+export class VVoid extends BaseValue {
+  readonly type = "Void";
+
+  constructor() {
+    super();
+  }
+  valueToString() {
+    return "()";
+  }
+}
+export const vVoid = () => new VVoid();
